Cursor Restructuring
I fell into treating the cursor like a player character and made all actions center on it. This worked fine until sync with the grid began which caused me to signal up to the game state.

I made the decision to restructure and make player input happen at the game state and call down to cursor to move. All the logic is happening in the game state (as it should) and cursor is merely a graphical element that moves on command.

Still looking to bound cursor within the grid. Logic is better structured to manage this

*Input delay was shortened a bit more.
*Game display was changed to stretch when widow is enlarged.
*Renamed Input-Delay in GameState.cs to Input_Is_Delay for better readability.
*Added readable sections to GameState.cs, better readability.