Bounded Cursor by the Power of Math
Cursor is now bounded to the grid according to the script defined grid size, and cursor will now reflect in the board state.

The Cursor's displayed position is pre-calculated and stored in the grid tile, and gets displayed in the corner to show the player where the cursor is within the isometric grid.

This is part of the transition to a better gamestate dependent system. In the future this game state will be used to save progress and manipulate the board/ communicate data.